+++ doc/changelog.rst
@@ -14,6 +14,15 @@ Patch version changes indicate backward compatible bug fixes.
 
 To install a specific version of the library you would run ``pip install mne-nirs==0.0.6``, where ``0.0.6`` is the version you wish to install.
 
+v0.1.2-development
+------
+
+Enhancements
+
+* Add :meth:`mne_nirs.channels.list_sources`. By `Robert Luke`_.
+* Add :meth:`mne_nirs.channels.list_detectors`. By `Robert Luke`_.
+
+
 v0.1.1
 ------